Laaarge Update (but they all are aren't they because I barely ever commit)
Added:
- New large seperator comments for each section (looks really good in
minimap)
- List of names of all known bots
- Smart ally system (shh)
- setTarget function which sets the target variable as the closest bot

Modified:
- Slight change to colours, also changes to green on win
- Slight change to radar, to stop it from essentially disabling our bot
when we don't have a target
- Made big modifications to onPaint, now shows us a triangle between
targets position, predicted position and our position
- Robot doesn't get stuck when hitting a wall it just changes direction.

* Git decided it didn't like me (most likely because of the 100 extra
lines near the start of the file) and the git diff of Feral.java is
terrible, don't
try reading it - its a mess :)

------------------

Also, made a new robot called Glue. This robot is essentially Walls on
roids.

It uses basically the same movement code as sample.Walls, rearranged
obviously, but has the targeting and firing of Feral.

It is lethal, and performs almost as well as Feral.

It's only real weakness is it's movement scheme, which is extremely
predictable, even by very basic targeting systems.